+++ tests/either.test.ts
@@ -171,6 +171,17 @@ describe('Either pure functions', () => {
     })
   })
 
+  describe('toMaybe', () => {
+    it('direct call', () => {
+      const left = Either.left<string, number>('error')
+      const right = Either.right<string, number>(144)
+      const nothing = Either.toMaybe(left)
+      const just = Either.toMaybe(right)
+      expect(nothing.toString()).toBe('Nothing()')
+      expect(just.toString()).toBe('Just(144)')
+    })
+  })
+
   describe('filter', () => {
     const predicate = (x: number) => x > 150
 
@@ -446,5 +457,18 @@ describe('Cases from docs', () => {
 
     expect(resultLeft).toBe(10)
     expect(resultRight).toBe(300)
+  })
+
+  it('toMaybe', () => {
+    const left = Either.left<string, number>('Some Error')
+    const right = Either.right<string, number>(150)
+    const rightVoid = Either.right<string, number | void>(undefined)
+
+    const resultLeft = left.toMaybe() // Nothing()
+    const resultRight = right.toMaybe() // Just(150)
+    const resultRightVoid = rightVoid.toMaybe() // Nothing()
+    expect(resultLeft.toString()).toBe('Nothing()')
+    expect(resultRight.toString()).toBe('Just(150)')
+    expect(resultRightVoid.toString()).toBe('Nothing()')
   })
 })