+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] :bomb:
 
+## :bulb: [0.3.0] - 2019-05-24
+
+### :gift: Added
+- `toEither` as pure function for `Maybe`
+- `toEither` as method for `Maybe`
+- `toMaybe` as pure function for `Either`
+- `toMaybe` as method for `Either`
+- `dark theme` for docs
+
+### :surfer: Changed
+- Fixed mistake in docs with link to interface `Maybe` or `Either` from pure function
+- Fixed mistakes in docs for calling pure functions in `Maybe` examples
+
 ## :bulb: [0.2.0] - 2019-05-21
 
 ### :gift: Added
@@ -19,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed mistake in naming `mather` to `matcher` :sweat_smile:
 - Fixed type `ApplicativeResult` (it could not return the nullable, now this will be available)
 - :fire: **WARNING** `MaybeShape` marked as **Deprecated**!
+- :fire: `Maybe.chain` was fixed, now will returns `Nothing` if the method was called on `Nothing`!
 
 ## :bulb: [0.1.0] - 2019-05-19
 
@@ -31,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Smoke test and really build to **Travis CI**.
 - Labels of `npm`, `coverage`, `travis`.
 
-[unreleased]: https://github.com/snatvb/monad-maniac/compare/v0.1.0...develop
-[0.1.0]: https://github.com/snatvb/monad-maniac/compare/v0.0.1...v0.1.0
+[unreleased]: https://github.com/snatvb/monad-maniac/compare/v0.3.0...develop
+[0.3.0]: https://github.com/snatvb/monad-maniac/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/snatvb/monad-maniac/compare/v0.1.0...v0.2.0
+[0.1.0]: https://github.com/snatvb/monad-maniac/compare/v0.0.1...v0.1.0